Rename and adjust class hierarchy for clarity
