+++ app/services/renewal_reminders/bulk_renewal_reminders_service_base.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+module RenewalReminders
+
+  class BulkRenewalRemindersServiceBase < WasteExemptionsEngine::BaseService
+    include WasteExemptionsEngine::ApplicationHelper
+
+    private
+
+    def template
+      raise NotImplementedError
+    end
+
+    def message_type
+      raise NotImplementedError
+    end
+
+    def all_active_exemptions_registration_ids
+      WasteExemptionsEngine::RegistrationExemption
+        .all_active_exemptions
+        .where(expires_on: expires_in_days.days.from_now.to_date)
+        .pluck(:registration_id)
+    end
+
+    def default_scope
+      WasteExemptionsEngine::Registration
+    end
+
+    def expires_in_days
+      raise(NotImplementedError)
+    end
+  end
+end
