+++ app/services/renewal_reminders/first_renewal_reminder_service.rb
@@ -2,7 +2,7 @@
 
 module RenewalReminders
 
-  class FirstRenewalReminderService < RenewalReminderEmailServiceBase
+  class FirstRenewalReminderService < BulkRenewalRemindersEmailService
     private
 
     def send_email(registration)
