+++ app/services/renewal_reminders/second_renewal_reminder_email_service.rb
@@ -4,7 +4,7 @@ require "notifications/client"
 
 module RenewalReminders
 
-  class SecondRenewalReminderEmailService < RenewalReminderEmailService
+  class SecondRenewalReminderEmailService < RenewalReminderEmailServiceBase
 
     # For CanHaveCommunicationLog
     def communications_log_params
