+++ app/services/renewal_reminders/temporary_first_renewal_reminder_service.rb
@@ -2,7 +2,7 @@
 
 module RenewalReminders
 
-  class TemporaryFirstRenewalReminderService < RenewalReminderEmailServiceBase
+  class TemporaryFirstRenewalReminderService < BulkRenewalRemindersEmailService
     private
 
     def send_email(registration)
