+++ app/services/renewal_reminders/temporary_second_renewal_reminder_service.rb
@@ -2,7 +2,7 @@
 
 module RenewalReminders
 
-  class TemporarySecondRenewalReminderService < RenewalReminderEmailServiceBase
+  class TemporarySecondRenewalReminderService < BulkRenewalRemindersEmailService
     private
 
     def send_email(registration)
